Split functionality out of __init__.py, re-enable pylint
